+++ changelog.md
@@ -2,10 +2,11 @@
 
 ## [5.18.1 (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v5.18.0...main)
 
-> \[!IMPORTANT\]
+> [!IMPORTANT]
 > This version is not released yet and is under active development.
 
 - \[mpm\] Drop support for Python 3.9.
+- \[mpm\] Run tests and actions on `ubuntu-24.04` instead of `ubuntu-22.04`.
 - \[mpm\] Replace local platform utilities by `extra-platforms` dependency.
 - \[mpm\] Add a Sankey diagram of all supported package managers.