+++ changelog.md
@@ -5,9 +5,12 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- \[mpm\] Add official support for Python 3.13.
 - \[mpm\] Drop support for Python 3.9.
-- \[mpm\] Run tests and actions on `ubuntu-24.04` instead of `ubuntu-22.04`.
 - \[mpm\] Replace local platform utilities by `extra-platforms` dependency.
+- \[mpm\] Run tests on Python 3.14-dev.
+- \[mpm\] Run tests and actions on `ubuntu-24.04` instead of `ubuntu-22.04`.
+- \[mpm\] Run tests on `macos-15` instead of `macos-14`.
 - \[mpm\] Add a Sankey diagram of all supported package managers.
 
 ## [5.18.0 (2024-08-02)](https://github.com/kdeldycke/meta-package-manager/compare/v5.17.0...v5.18.0)